refactored code into FormulaInputRangeSyncFeature
