+++ packages/ag-grid-enterprise/src/widgets/formulaRangeUtils.ts
@@ -1,0 +1,145 @@
+import type { BeanCollection, CellRange } from 'ag-grid-community';
+
+// Allow partial ranges (eg "A1:") so we keep typing within the same token until a breaking operator is entered.
+export const CELL_OR_RANGE_REGEX = /\$?[A-Za-z]+\$?[0-9]+(?::\$?[A-Za-z]+\$?[0-9]+)?:?/g;
+
+// Token/range color helpers
+export const getTokenColorClass = (colorIndex: number): string => `ag-formula-token-color-${colorIndex + 1}`;
+export const getRangeColorClass = (colorIndex: number): string => `ag-formula-range-color-${colorIndex + 1}`;
+
+// Keep token and range overlay classes in sync for a given color index.
+export const getColorClassesForRef = (
+    _ref: string,
+    colorIndexOverride?: number | null
+): { tokenClass: string; rangeClass: string; colorIndex: number } => {
+    const index = colorIndexOverride ?? 0;
+
+    return {
+        tokenClass: getTokenColorClass(index),
+        rangeClass: getRangeColorClass(index),
+        colorIndex: index,
+    };
+};
+
+// Range overlay helpers
+export const getRangeColorIndexFromClass = (colorClass?: string | null): number | null => {
+    if (!colorClass) {
+        return null;
+    }
+
+    const match = /ag-formula-range-color-(\d+)/.exec(colorClass);
+
+    if (!match) {
+        return null;
+    }
+
+    const parsed = parseInt(match[1], 10);
+    return Number.isFinite(parsed) ? parsed - 1 : null;
+};
+
+export const tagRangeWithFormulaColor = (
+    range: CellRange | undefined,
+    ref: string,
+    colorIndex?: number | null
+): void => {
+    if (!range) {
+        return;
+    }
+
+    const { rangeClass } = getColorClassesForRef(ref, colorIndex);
+    range.colorClass = rangeClass;
+};
+
+// Range helpers
+export const getCellRangeParams = (beans: BeanCollection, ref: string) => {
+    const match = /^\$?([A-Za-z]+)\$?(\d+)(?::\$?([A-Za-z]+)\$?(\d+))?$/.exec(ref);
+    if (!match) {
+        return null;
+    }
+
+    const { formula } = beans;
+
+    const [, startColRef, startRowStr, endColRef, endRowStr] = match;
+    const startCol = formula?.getColByRef(startColRef);
+    const endCol = formula?.getColByRef(endColRef ?? startColRef);
+
+    if (!startCol || !endCol) {
+        return null;
+    }
+
+    const rowStartIndex = parseInt(startRowStr, 10) - 1;
+    const rowEndIndex = endRowStr ? parseInt(endRowStr, 10) - 1 : rowStartIndex;
+
+    return {
+        rowStartIndex,
+        rowEndIndex,
+        columnStart: startCol,
+        columnEnd: endCol,
+    };
+};
+
+export const getLatestRangeRef = (beans: BeanCollection): string | null => {
+    const ranges = beans.rangeSvc?.getCellRanges();
+    const latest = ranges?.length ? ranges[ranges.length - 1] : null;
+
+    if (!latest) {
+        return null;
+    }
+
+    return rangeToRef(beans, latest);
+};
+
+export const rangeToRef = (beans: BeanCollection, range: CellRange): string | null => {
+    const { rangeSvc, formula } = beans;
+
+    if (!rangeSvc || !formula) {
+        return null;
+    }
+
+    const startRow = rangeSvc.getRangeStartRow(range);
+    const endRow = rangeSvc.getRangeEndRow(range);
+
+    if (!startRow || !endRow || startRow.rowPinned || endRow.rowPinned) {
+        return null;
+    }
+
+    const rowStartIndex = Math.min(startRow.rowIndex!, endRow.rowIndex!) + 1;
+    const rowEndIndex = Math.max(startRow.rowIndex!, endRow.rowIndex!) + 1;
+
+    const columns = range.columns;
+
+    if (!columns?.length) {
+        return null;
+    }
+
+    const sorted = [...columns];
+    const startCol = sorted[0];
+    const endCol = sorted[sorted.length - 1];
+
+    const colStartRef = formula.getColRef(startCol as any);
+    const colEndRef = formula.getColRef(endCol as any);
+
+    if (!colStartRef || !colEndRef) {
+        return null;
+    }
+
+    const sameCol = colStartRef === colEndRef;
+    const sameRow = rowStartIndex === rowEndIndex;
+
+    if (sameCol && sameRow) {
+        return `${colStartRef}${rowStartIndex}`;
+    }
+
+    return `${colStartRef}${rowStartIndex}:${colEndRef}${rowEndIndex}`;
+};
+
+export const getRefsFromText = (text: string): Set<string> => {
+    // Extract all A1-style refs/ranges from raw text to keep grid ranges in sync.
+    const refs = new Set<string>();
+    let match: RegExpExecArray | null;
+    CELL_OR_RANGE_REGEX.lastIndex = 0;
+    while ((match = CELL_OR_RANGE_REGEX.exec(text)) != null) {
+        refs.add(match[0]);
+    }
+    return refs;
+};
